+++ qsdsan/_sanunit.py
@@ -41,11 +41,12 @@ class SanUnit(Unit, isabstract=True):
     uptime_ratio : float
         Uptime of the unit to adjust add_OPEX, should be in [0,1]
         (i.e., a unit that is always operating has an uptime_ratio of 1).
-    lifetime : float
-        Lifetime of this unit in year.
+    lifetime : float or dict
+        Lifetime of this unit (float) or individual equipment within this unit
+        (dict) in year.
         It will be used to adjust cost and emission calculation in TEA and LCA.
-        If left as None, its lifetime will be assumed to be the same as the
-        TEA/LCA lifetime.
+        Equipment without provided lifetime will be assumed to have the same
+        lifetime as the TEA/LCA.         
     
     See Also
     --------
@@ -56,7 +57,8 @@ class SanUnit(Unit, isabstract=True):
     _stacklevel = 7
     ticket_name = 'SU'
 
-    def __init__(self, ID='', ins=None, outs=(), thermo=None, **kwargs):
+    def __init__(self, ID='', ins=None, outs=(), thermo=None, 
+                 equipments=(), **kwargs):
         self._register(ID)
         self._specification = None
         self._load_thermo(thermo)
@@ -65,6 +67,9 @@ class SanUnit(Unit, isabstract=True):
         self._init_utils()
         self._init_results()
         self._assert_compatible_property_package()
+        self.equipments = equipments
+        for equip in equipments:
+            equip._linked_unit = self
         for attr, val in kwargs.items():
             setattr(self, attr, val)
 
@@ -144,7 +149,21 @@ class SanUnit(Unit, isabstract=True):
         '''Print information of the unit, including waste stream-specific information.'''
         print(self._info(T, P, flow, composition, N, IDs, stream_info))
     
-    def add_construction(self, add_unit=True, add_design=True, add_cost=True):
+    def add_equipment_design(self):
+        for equip in self.equipments:
+            name = equip.name
+            self.design_results[name] = equip._design(self)
+            self._units.update(equip.design_units)
+            self._BM[name] = equip.BM
+            if equip.lifetime:
+                self._equipment_lifetime[name] = equip.lifetime
+    
+    def add_equipment_cost(self):
+        for equip in self.equipments:
+            self.purchase_cost[equip] = equip._cost(self)
+    
+    def add_construction(self, add_unit=True, add_design=True, add_cost=True,
+                         add_lifetime=True):
         '''Batch-adding construction unit, designs, and costs.'''
         for i in self.construction:
             if add_unit:
@@ -152,7 +171,9 @@ class SanUnit(Unit, isabstract=True):
             if add_design:
                 self.design_results[i.item.ID] = i.quantity
             if add_cost:
-                self.purchase_costs[i.item.ID] = i.cost 
+                self.purchase_costs[i.item.ID] = i.cost
+            if add_lifetime and i.lifetime:
+                self._equipment_lifetime[i.item.ID] = i.lifetime             
     
     @property
     def components(self):
@@ -240,12 +261,18 @@ class SanUnit(Unit, isabstract=True):
     @property
     def lifetime(self):
         '''
-        [float] Lifetime of this unit in year.
+        [float] or [dict] Lifetime of this unit (float) or individual equipment
+        within this unit (dict) in year.
         It will be used to adjust cost and emission calculation in TEA and LCA.
-        If left as None, its lifetime will be assumed to be the same as the
-        TEA/LCA lifetime.
+        Equipment without provided lifetime will be assumed to have the same
+        lifetime as the TEA/LCA. 
         '''
-        return self._lifetime
+        if self._lifetime is not None:
+            return self._lifetime
+        elif self._equipment_lifetime:
+            return self._equipment_lifetime
+        else:
+            return None
     @lifetime.setter
     def lifetime(self, i):
         self._lifetime = float(i)